Feat/user routes (#7)
* just testing...

* pre-commit

* index.html

* index.html

* Create docs.yaml

* workflow test

* pre-commit

* Update docs.yaml

* name

* typo

* install swag

* install swag

* install swag correctly

* typo

* km

* cd

* file path

* chdir

* auto commit

* auto commit

* alligining

* commt

* change

* ""

* docs: update swagger docs

* feat: implement user route, sercive, repo

* Pre-commit

* feat: update user routes and implement pagination for GetAllUsers

* feat: add task retrieval methods for user and group with pagination

* pre-commit

* feat: enhance transaction handling and pagination defaults across services and routes

* pre-commit

* Changes in regards to pr

* fix: update ValidateSession to return structured response on transaction error

* pre-commit

---------

+++ package/service/session_service.go
@@ -9,6 +9,7 @@ import (
 	"github.com/mini-maxit/backend/package/domain/models"
 	"github.com/mini-maxit/backend/package/domain/schemas"
 	"github.com/mini-maxit/backend/package/repository"
+	"github.com/mini-maxit/backend/package/utils"
 	"gorm.io/gorm"
 )
 
@@ -52,6 +53,8 @@ func (s *SessionServiceImpl) CreateSession(tx *gorm.DB, userId int64) (*schemas.
 			return nil, tx.Error
 		}
 	}
+
+	defer utils.TransactionPanicRecover(tx)
 
 	_, err := s.userRepository.GetUser(tx, userId)
 	if err != nil {
@@ -105,6 +108,11 @@ func (s *SessionServiceImpl) CreateSession(tx *gorm.DB, userId int64) (*schemas.
 
 func (s *SessionServiceImpl) ValidateSession(sessionId string) (schemas.ValidateSessionResponse, error) {
 	tx := s.database.Connect().Begin()
+	if tx.Error != nil {
+		return schemas.ValidateSessionResponse{Valid: false, UserId: -1}, tx.Error
+	}
+
+	defer utils.TransactionPanicRecover(tx)
 
 	session, err := s.sessionRepository.GetSession(tx, sessionId)
 	if err != nil {
@@ -155,6 +163,11 @@ func (s *SessionServiceImpl) RefreshSession(sessionId string) (*schemas.Session,
 
 func (s *SessionServiceImpl) InvalidateSession(sessionId string) error {
 	tx := s.database.Connect().Begin()
+	if tx.Error != nil {
+		return tx.Error
+	}
+
+	defer utils.TransactionPanicRecover(tx)
 
 	err := s.sessionRepository.DeleteSession(tx, sessionId)
 	if err != nil {
